backend/scraping/license_scraper.py
"""
License verification scraper for state medical boards.
"""

import aiohttp
from bs4 import BeautifulSoup
from typing import Dict, Optional, Any
import logging
import re

logger = logging.getLogger(__name__)


async def verify_license(license_number: str, state: str, provider_name: str = "") -> Dict[str, Any]:
    """
    Verify medical license with state board (placeholder implementation).
    
    Note: Actual implementation would need to integrate with each state's
    medical board website, which varies by state.
    
    Args:
        license_number: License number
        state: State code (2 letters)
        provider_name: Provider name for verification
        
    Returns:
        Dictionary with verification results
    """
    # This is a placeholder - actual implementation would need state-specific logic
    state = state.upper()
    
    # State board URLs (examples - would need to be expanded)
    state_boards = {
        "CA": "https://www.mbc.ca.gov/breeze/license_lookup.php",
        "TX": "https://www.tmb.state.tx.us/page/verify-a-license",
        "NY": "https://www.op.nysed.gov/prof/med/medlic.htm",
        "FL": "https://www.flhealthsource.gov/mqa/",
    }
    
    if state not in state_boards:
        return {
            "verified": False,
            "error": f"State {state} not yet supported",
            "status": "unknown"
        }
    
    # Placeholder implementation
    # In production, this would:
    # 1. Navigate to state board website
    # 2. Fill in license number and name
    # 3. Parse results
    # 4. Return verification status
    
    try:
        # For now, return a mock response
        # In production, use Selenium or similar for dynamic sites
        return {
            "verified": True,  # Placeholder
            "status": "active",  # Placeholder
            "issue_date": None,
            "expiry_date": None,
            "board_url": state_boards[state],
            "note": "Placeholder implementation - requires state-specific integration"
        }
    except Exception as e:
        logger.error(f"Error verifying license: {e}")
        return {
            "verified": False,
            "error": str(e),
            "status": "unknown"
        }


async def scrape_state_board(license_number: str, state: str, provider_name: str = "") -> Dict[str, Any]:
    """
    Scrape state medical board website for license information.
    
    Args:
        license_number: License number
        state: State code
        provider_name: Provider name
        
    Returns:
        License information dictionary
    """
    # This would use Selenium for dynamic sites or aiohttp for static sites
    # Implementation varies by state
    
    return await verify_license(license_number, state, provider_name)

