requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlite3  # Built-in, but listed for clarity

# Data processing
pandas==2.1.3
numpy==1.26.2

# HTTP client
aiohttp==3.9.1

# Web scraping
beautifulsoup4==4.12.2
selenium==4.15.2

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Phone number validation
phonenumbers==8.13.25

# Pydantic for data validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Logging
python-json-logger==2.0.7

# Environment variables
python-dotenv==1.0.0

